refactor: load model and tokenizer as default
